Fix: 피드백 기반 수정

+++ Source/Endless_Hotel/Anomaly/Object/EightExit/Light/Anomaly_Object_Light.h
@@ -32,7 +32,7 @@ public:
 
 private:
 	UPROPERTY(EditAnywhere, Category = "Index")
-	int32 LightIndex = 0;
+	int8 LightIndex = 0;
 
 	UPROPERTY(EditAnywhere)
 	TObjectPtr<class UPointLightComponent> PointLight;
